Added checks for ownership and existence

+++ output/opendap_post2.sh
@@ -130,7 +130,7 @@ for server in ${SERVERS[*]}; do
       cat run.properties >> $RUNPROPERTIES
       rm run.properties # so we don't keep appending to it
       $SCRIPTDIR/metadata.pl --redact --jsonify --metadatafile $RUNPROPERTIES --converted-file-name asgs.instance.status.json 2>> $SYSLOG
-      sed --in-place "s/$USER/\$USER/g" asgs.instance.status.json 2>> $SYSLOG 
+      sed --in-place "s/$USER/\$USER/g" asgs.instance.status.json 2>> $SYSLOG
    fi
    # FIXME: enable Operator to override TDS parameter settings from platforms.sh
    _THIS="output/opendap_post2.sh-->$server"
@@ -292,7 +292,7 @@ wget $DOWNLOADPREFIX/$STORMNAMEPATH/$OPENDAPSUFFIX/$logfile
 
 END
       $SCRIPTDIR/metadata.pl --jsonify --redact --metadatafile $RUNPROPERTIES --converted-file-name asgs.instance.status.json 2>> $SYSLOG
-      sed --in-place "s/$USER/\$USER/g" asgs.instance.status.json 2>> $SYSLOG 
+      sed --in-place "s/$USER/\$USER/g" asgs.instance.status.json 2>> $SYSLOG
    else
 cat <<END > ${SCENARIODIR}/opendap_results_notify_${server}.txt
 
@@ -444,7 +444,9 @@ SSHCMD
 # this block will be executed on the remote server,
 # variables are interpolated locally unless escaped
 # with a backslash, '\'
-ln -s "$OPENDAPBASEDIR/$STORMNAMEPATH" "$OPENDAPBASEDIR/$ALTSTORMNAMEPATH"
+if [[ -e "$OPENDAPBASEDIR/$STORMNAMEPATH" && ! -e "$OPENDAPBASEDIR/$ALTSTORMNAMEPATH" ]]; then
+   ln -s "$OPENDAPBASEDIR/$STORMNAMEPATH" "$OPENDAPBASEDIR/$ALTSTORMNAMEPATH"
+fi
 SSHCMD
             if [[ $? != 0 ]]; then
                MSG="$SCENARIO: $_THIS: Failed to create symbolic link for the storm name."
@@ -511,7 +513,7 @@ SSHCMD
               unset MSG
             else
               opendapEmailSent=yes
-            fi 
+            fi
             fileIndex=`expr $fileIndex + 1` 2>> $SCENARIOLOG
             continue
          fi
@@ -696,7 +698,7 @@ SSHCMD
               unset MSG
             else
               opendapEmailSent=yes
-            fi 
+            fi
             continue
          fi
          chmod +r "$file" 2>> $SYSLOG
@@ -738,7 +740,9 @@ SSHCMD
       while [[ $partialPath != $OPENDAPBASEDIR  ]]; do
          retry=0
          while [[ $retry -lt $timeoutRetryLimit ]]; do
-            chmod a+wx $partialPath 2>> $SYSLOG
+            if [[ $(stat -c %u "$partialPath") -eq $(id -u $USER) ]]; then
+               chmod a+wx $partialPath 2>> $SYSLOG
+            fi
             if [[ $? != 0 ]]; then
                MSG="$SCENARIO: $_THIS: Failed to change permissions on the directory ${partialPath}."
                if [ "$MANUAL" == 1 ]; then
@@ -765,7 +769,7 @@ SSHCMD
       done
       #
       # create link with storm name instead of storm number
-      if [[ $TROPICALCYCLONE != off ]]; then
+      if [[ $TROPICALCYCLONE != off && -e "$OPENDAPBASEDIR/$STORMNAMEPATH" && ! -e "$OPENDAPBASEDIR/$ALTSTORMNAMEPATH" ]]; then
          ln -s $OPENDAPBASEDIR/$STORMNAMEPATH $OPENDAPBASEDIR/$ALTSTORMNAMEPATH 2>> $SYSLOG
       fi
       for file in ${FILES[*]}; do
@@ -789,7 +793,7 @@ SSHCMD
               unset MSG
             else
               opendapEmailSent=yes
-            fi 
+            fi
             continue
          fi
          chmod +r $file 2>> $SYSLOG
@@ -848,6 +852,6 @@ SSHCMD
           warn "$MSG"
         fi
         unset MSG
-      fi 
+      fi
    fi
 done # end loop over opendap servers
